Remove interface LookupJoinPageBuilder
